// src/api/models/ContactoDirectoResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContactoDirectoResponse = {
    id?: number;
    clienteId?: number;
    tipoContacto?: string | null;
    telefono?: string | null;
    email?: string | null;
    principal?: boolean;
    activo?: boolean;
    fechaCreacion?: string;
    fechaActualizacion?: string | null;
    readonly tipoContactoDescripcion?: string | null;
    readonly tieneTelefono?: boolean;
    readonly tieneEmail?: boolean;
};