// src/api/models/ContactoEmpresaDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContactoEmpresaDto = {
    id?: number;
    clienteEmpresaId?: number;
    nombre?: string | null;
    cargo?: string | null;
    telefono?: string | null;
    email?: string | null;
    activo?: boolean;
    fechaCreacion?: string;
    fechaActualizacion?: string | null;
};