// src/api/models/ContactosEmpresaRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContactosEmpresaRequest = {
    nombre: string;
    cargo?: string | null;
    email?: string | null;
    telefono?: string | null;
    readonly esValido?: boolean;
};